+++ include/studium/input.hpp
@@ -8,12 +8,26 @@ namespace studium
     //! Wraps all functions and structures related to input handling.
     namespace input
     {
-        // Input callbacks, you shouldn't need to care about those
-        // in general.
+        // Input callbacks
+
+        //! Callback function for keyboard input events.
+        //  \note This function should not be called directly.
         void keyboard_callback(GLFWwindow*, int, int, int, int);
+        
+        //! Callback function for mouse repositioning events.
+        //  \note This function should not be called directly.
         void mouse_pos_callback(GLFWwindow*, double, double);
+
+        //! Callback function for mouse click events.
+        //  \note This function should not be called directly.
         void mouse_btn_callback(GLFWwindow*, int, int, int);
+
+        //! Callback function for mouse scroll wheel events.
+        //  \note This function should not be called directly.
         void mouse_scl_callback(GLFWwindow*, double, double);
+
+        //! Callback function for gamepad events.
+        //  \note This function should not be called directly.
         void gamepad_callback(int, int);
     }
 }