+++ include/studium/render.hpp
@@ -21,8 +21,21 @@ namespace studium
             glm::vec2   size;
         public:
             ~Renderer2D();
+            /*!
+             * Initializes both the window and the rendering context.
+             * If already initialized, does nothing.
+             * \return Whether the initialization process succeeded.
+             */  
             bool init();
+
+            /*!
+             * Performs an infinite game loop, while the game doesn't trigger an exit signal
+             * to the window.
+             * \return Whether the attempt to run a game loop succeeded.
+             */
             bool game_loop();
+
+            //!  Performs a full disposal of the window and the rendering context.
             void unload();
         };
     }