+++ src/engine.cpp
@@ -1,0 +1,7 @@
+#include <studium/engine.hpp>
+
+const char*
+studium::engine::get_version_string()
+{
+    return STUDIUM_VERSION_STRING;
+}